fix Pool.Put(nil)
Interfaces are compound types consisting of a runtime type and a value.
Only if both components are nil is the interface value itself nil. So as
soon as a concrete value is converted into an interface value that
interface value is != nil.

    package main

    import (
            "fmt"
            "net/url"
    )

    func main() {
            var c *url.URL         // c is a concrete value
            var i fmt.Stringer = c // i is an interface value

            fmt.Println(c, c == nil)
            fmt.Println(i, i == nil)
    }

    # Output:
    # <nil> true
    # <nil> false

This happened in Pool.Put, where a *Disque was converted into a
"github.com/youtube/vitess/go/pools".Resource. Vitesses ResourcePool
then did not recognize this as a nil value and would eventually return
it from a Get call. At that point, Pool.Get converts the nil back into a
value of type *Disque just waiting to produce a nil pointer panic.